Move common permissions types to different utility file
