+++ mathesar_ui/src/systems/permissions/index.ts
@@ -1,4 +1,4 @@
 export { default as PermissionsModal } from './PermissionsModal.svelte';
 export { default as PermissionsOverview } from './overview/PermissionsOverview.svelte';
 export { default as TransferOwnership } from './transfer-ownership/TransferOwnership.svelte';
-export { type AccessControlConfig } from './overview/utils';
+export { type AccessControlConfig } from './overview/overviewUtils';
