+++ mathesar_ui/src/systems/permissions/overview/Owner.svelte
@@ -4,9 +4,10 @@
   import type { Role } from '@mathesar/models/Role';
   import type { ImmutableMap } from '@mathesar-component-library';
 
+  import type { PermissionsMetaData } from '../permissionsUtils';
+
   import OverviewSection from './OverviewSection.svelte';
   import RoleWithChildren from './RoleWithChildren.svelte';
-  import type { PermissionsMetaData } from './utils';
 
   type Privilege = $$Generic;
