+++ mathesar_ui/src/systems/permissions/permissionsUtils.ts
@@ -4,25 +4,9 @@ import type { Role } from '@mathesar/models/Role';
 import type { AsyncStoreValue } from '@mathesar/stores/AsyncStore';
 import type { ImmutableMap } from '@mathesar-component-library';
 
-import {
-  type AccessLevelConfig,
-  RoleAccessLevelAndPrivileges,
-} from './RoleAccessLevelAndPrivileges';
-
 export interface RolePrivileges<Privilege> {
   role_oid: Role['oid'];
   direct: Privilege[];
-}
-
-export interface AccessControlConfig<AccessLevel, Privilege> {
-  allPrivileges: {
-    id: Privilege;
-    help: string;
-  }[];
-  access: {
-    levels: readonly AccessLevelConfig<AccessLevel, Privilege>[];
-    default: AccessLevel;
-  };
 }
 
 export interface PermissionsMetaData<Privilege> {
@@ -40,19 +24,4 @@ export interface PermissionsAsyncStores<Privilege, E = string> {
   permissionsMetaData: Readable<
     AsyncStoreValue<PermissionsMetaData<Privilege>, E>
   >;
-}
-
-export function getObjectAccessPrivilegeMap<A, P>(
-  accessLevelConfig: readonly AccessLevelConfig<A, P>[],
-  privilegesForRole: ImmutableMap<Role['oid'], RolePrivileges<P>>,
-) {
-  return privilegesForRole.mapValues(
-    (entry) =>
-      new RoleAccessLevelAndPrivileges({
-        roleOid: entry.role_oid,
-        accessLevelConfig,
-        privileges: [...entry.direct],
-        savedPrivileges: [...entry.direct],
-      }),
-  );
 }
